requirements.txt
streamlit 
google-generativeai
python-dotenv
pdf2image #Convert PDF to image

